feat: New Footer (fix #2518)
