+++ src/amo/components/Footer/styles.scss
@@ -1,0 +1,49 @@
+@import "~core/css/inc/mixins";
+@import "~amo/css/inc/vars";
+@import "~ui/css/vars";
+
+.Footer {
+  background: $footer-base-color;
+  color: $white;
+  padding: 20px;
+
+  @include respond-to(medium) {
+    display: flex;
+    justify-content: space-between;
+  }
+}
+
+.Footer-language-and-links {
+  max-width: 220px;
+  width: 100%;
+}
+
+.Footer-links {
+  list-style: none;
+  margin: 20px 0 0;
+  padding: 0;
+}
+
+.Footer-link,
+.Footer-link:link {
+  color: $white;
+  display: block;
+  margin: 0 0 10px;
+  text-decoration: none;
+
+  &:active,
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.Footer-mozilla-link {
+  display: inline-block;
+  margin-top: 10px;
+
+  @include respond-to(medium) {
+    @include margin-end(10px);
+
+    margin-top: auto;
+  }
+}
